refactor: split MA library types.rs and lib.rs into focused modules
